requirements2.txt
streamlit
pandas
matplotlib
google-generativeai

